+++ packages/headless/src/commerce.index.ts
@@ -21,7 +21,7 @@ export type {LogLevel, LoggerOptions} from './app/logger';
 export type {ProductRecommendation} from './api/search/search/product-recommendation';
 
 // Actions
-export * from './features/product-listing/v2/product-listing-v2-actions-loader';
+export * from './features/commerce/product-listing/product-listing-actions-loader';
 export * from './features/configuration/configuration-actions-loader';
 export * from './features/analytics/search-analytics-actions-loader';
