+++ packages/headless/src/controllers/commerce/product-listing/pager/headless-product-listing-pager.ts
@@ -1,5 +1,5 @@
 import {CommerceEngine} from '../../../../app/commerce-engine/commerce-engine';
-import {fetchProductListing} from '../../../../features/product-listing/v2/product-listing-v2-actions';
+import {fetchProductListing} from '../../../../features/commerce/product-listing/product-listing-actions';
 import {
   buildCorePager,
   PagerInitialState,
