+++ packages/headless/src/features/commerce/product-listing/product-listing-actions.ts
@@ -26,7 +26,7 @@ import {
 } from '../../analytics/analytics-utils';
 import {getFacetRequests} from '../../facets/generic/interfaces/generic-facet-request';
 import {logQueryError} from '../../search/search-analytics-actions';
-import {logProductListingV2Load} from './product-listing-v2-analytics';
+import {logProductListingV2Load} from './product-listing-analytics';
 
 export interface SetProductListingUrlPayload {
   /**
