+++ packages/headless/src/features/pagination/pagination-slice.test.ts
@@ -3,6 +3,7 @@ import {buildFetchProductListingResponse} from '../../test/mock-product-listing'
 import {buildFetchProductListingV2Response} from '../../test/mock-product-listing-v2';
 import {buildMockSearch} from '../../test/mock-search';
 import {deselectAllBreadcrumbs} from '../breadcrumb/breadcrumb-actions';
+import {fetchProductListing as fetchProductListingV2} from '../commerce/product-listing/product-listing-actions';
 import {toggleSelectAutomaticFacetValue} from '../facets/automatic-facet-set/automatic-facet-set-actions';
 import {
   deselectAllCategoryFacetValues,
@@ -23,7 +24,6 @@ import {
 import {change} from '../history/history-actions';
 import {getHistoryInitialState} from '../history/history-state';
 import {fetchProductListing} from '../product-listing/product-listing-actions';
-import {fetchProductListing as fetchProductListingV2} from '../product-listing/v2/product-listing-v2-actions';
 import {logSearchboxSubmit} from '../query/query-analytics-actions';
 import {restoreSearchParameters} from '../search-parameters/search-parameter-actions';
 import {executeSearch} from '../search/search-actions';
