+++ packages/headless/src/features/pagination/pagination-slice.ts
@@ -1,5 +1,6 @@
 import {createReducer} from '@reduxjs/toolkit';
 import {deselectAllBreadcrumbs} from '../breadcrumb/breadcrumb-actions';
+import {fetchProductListing as fetchProductListingV2} from '../commerce/product-listing/product-listing-actions';
 import {toggleSelectAutomaticFacetValue} from '../facets/automatic-facet-set/automatic-facet-set-actions';
 import {
   deselectAllCategoryFacetValues,
@@ -17,7 +18,6 @@ import {
 } from '../facets/range-facets/numeric-facet-set/numeric-facet-actions';
 import {change} from '../history/history-actions';
 import {fetchProductListing} from '../product-listing/product-listing-actions';
-import {fetchProductListing as fetchProductListingV2} from '../product-listing/v2/product-listing-v2-actions';
 import {restoreSearchParameters} from '../search-parameters/search-parameter-actions';
 import {executeSearch} from '../search/search-actions';
 import {updateActiveTab} from '../tab-set/tab-set-actions';
