+++ packages/headless/src/test/mock-commerce-state.ts
@@ -1,3 +1,4 @@
+import {getProductListingV2InitialState} from '../features/commerce/product-listing/product-listing-state';
 import {getConfigurationInitialState} from '../features/configuration/configuration-state';
 import {getCategoryFacetSetInitialState} from '../features/facets/category-facet-set/category-facet-set-state';
 import {getFacetOrderInitialState} from '../features/facets/facet-order/facet-order-state';
@@ -5,7 +6,6 @@ import {getFacetSetInitialState} from '../features/facets/facet-set/facet-set-st
 import {getDateFacetSetInitialState} from '../features/facets/range-facets/date-facet-set/date-facet-set-state';
 import {getNumericFacetSetInitialState} from '../features/facets/range-facets/numeric-facet-set/numeric-facet-set-state';
 import {getPaginationInitialState} from '../features/pagination/pagination-state';
-import {getProductListingV2InitialState} from '../features/product-listing/v2/product-listing-v2-state';
 import {getSortInitialState} from '../features/sort/sort-state';
 import {CommerceAppState} from '../state/commerce-app-state';
